feat: lock/unlock project

+++ api/src/api/modules/organization/services/project_service.py
@@ -39,6 +39,7 @@ class ProjectService:
         project_data = {
             "name": project.name,
             "description": project.description,
+            "locked": False,
             "location": dir_path,
             "created_at": datetime.now(),
             "updated_at": datetime.now(),
@@ -57,6 +58,10 @@ class ProjectService:
         new_project = self.projects_storage.find_one({"name": project_name})
         if new_project is None:
             raise BadRequestException(f"Project with name {project_name} doesn’t exist.")
+        
+        if self.is_project_locked(project_name):
+            raise BadRequestException(f"Project with name {project_name} is locked.")
+        
 
         new_project["name"] = project.name if project.name else new_project["name"]
         new_project["description"] = (
@@ -89,5 +94,32 @@ class ProjectService:
         if not self.projects_storage.exists({"name": project_name}):
             raise BadRequestException(f"Project with name {project_name} doesn’t exist.")
 
+        if self.is_project_locked(project_name):
+            raise BadRequestException(f"Project with name {project_name} is locked.")
+
         self.file_management.delete_directory(project_name)
         self.projects_storage.delete_one({"name": project_name})
+
+    def lock_project(self, project_name: str) -> ProjectRes:
+        project = self.projects_storage.find_one({"name": project_name})
+        if project is None:
+            raise BadRequestException(f"Project with name {project_name} doesn’t exist.")
+        
+        project["locked"] = True
+        self.projects_storage.update({"name": project_name}, project)
+        return ProjectRes(**project)
+    
+    def unlock_project(self, project_name: str) -> ProjectRes:
+        project = self.projects_storage.find_one({"name": project_name})
+        if project is None:
+            raise BadRequestException(f"Project with name {project_name} doesn’t exist.")
+        
+        project["locked"] = False
+        self.projects_storage.update({"name": project_name}, project)
+        return ProjectRes(**project)
+
+    def is_project_locked(self, project_name: str) -> bool:
+        project = self.projects_storage.find_one({"name": project_name})
+        if project is None:
+            raise BadRequestException(f"Project with name {project_name} doesn’t exist.")
+        return project["locked"]
